Added test harness for the display abstraction that can generate SVG files for testing the behaviour of the renderer. No automation or assertions are linked up to this at the moment.
Changed the behaviour of the renderer to try and reduce the number of drawing operations that are performed. Rather than rendering the border grid as 4 separate lines for each square, we now render the grid in one pass, and then only use update operations to fill in the square colour/render the text as required. This should hopefully render faster on the eink screen rather than sending 120ish commands.

Joke is, we could just hard code the coordinates and make it a display-specific thing and not do all this math, but I am curious.