So turns out that insomnia means that you end up rewriting your morse parser at 2am, and then spending another 3 hours rewriting unit tests because apparently you were drunk the first time you wrote them and wrote entirely USELESS unit tests that were entirely unmaintainable, past Aidan, you are an idiot.
Morse seems to work on the STM32 now according to the debugger, still NFI why the display isn't working.

Enabled QuadSPI in the HAL as well as it's likely going to be used for external NOR flash storage for loading fonts and the wordle dictionary from, but that is a whole other world of pain to be encountered in the future.